feature-6/ fix typo in package name. rename exception
