+++ src/main/java/jewellery/inventory/exception/MappingException.java
@@ -1,0 +1,7 @@
+package jewellery.inventory.exception;
+
+public class MappingException extends RuntimeException {
+  public MappingException(Object resourceRequestDto) {
+    super(String.format("Can't map: %s", resourceRequestDto));
+  }
+}
