--- src/main/java/jewellery/inventory/exeption/ApiRequestException.java
@@ -1,12 +1,0 @@
-package jewellery.inventory.exeption;
-
-public class ApiRequestException extends RuntimeException {
-    public ApiRequestException(String message) {
-        super(message);
-    }
-
-    public ApiRequestException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
