--- src/main/java/jewellery/inventory/exeption/MappingException.java
@@ -1,11 +1,0 @@
-package jewellery.inventory.exeption;
-
-public class MappingException extends RuntimeException {
-    public MappingException(String message) {
-        super(message);
-    }
-
-    public MappingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
